apps/api/requirements.txt
# Volo API — AI Life Operating System Backend
# Python 3.11+

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# AI / LLM
anthropic>=0.40.0
openai>=1.12.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
pgvector>=0.2.5

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
cryptography>=42.0.0

# Real-time
python-socketio>=5.11.0
websockets>=12.0

# HTTP
httpx>=0.27.0
aiohttp>=3.9.0

# Task scheduling
apscheduler>=3.10.0

# Utilities
orjson>=3.9.0
python-multipart>=0.0.9
